Move Devices container file up into Simulator dir
